+++ pmset-sleeps.sh
@@ -1,0 +1,2 @@
+#!/bin/sh
+pmset -a displaysleep 30 disksleep 30 sleep 60 womp 1